requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.7
daphne==4.1.2
dj-database-url==2.2.0
Django==4.2.13
django-allauth==0.63.5
django-appconf==1.0.6
django-cors-headers==4.3.1
django-heroku==0.3.1
django-jazzmin==3.0.0
django-on-heroku==1.1.2
django-phonenumber-field==7.3.0
django-select2==8.1.2
django-twilio==0.14.3.3
djangorestframework==3.15.1
frozenlist==1.4.1
geoip2==4.8.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
maxminddb==2.6.2
multidict==6.0.5
phonenumbers==8.13.37
pillow==10.3.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
python-decouple==3.8
requests==2.32.3
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.0
twilio==9.1.0
Twisted==24.3.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
whitenoise==6.6.0
yarl==1.9.4
zope.interface==6.4.post2
